+++ beam/src/Rhyolite/DB/Beam.hs
@@ -1,0 +1,13 @@
+{-# Language FlexibleContexts #-}
+{-# Language TypeFamilies #-}
+-- | Utility functions for using Beam with Postgres
+module Rhyolite.DB.Beam where
+
+import Database.Beam
+import Database.Beam.Postgres
+
+import Database.PostgreSQL.Simple.Transaction
+
+-- | Run beam SQL statements inside a Postgres Serializable Transaction
+withTransactionSerializableRunBeamPostgres :: (MonadIO m) => Connection -> Pg a -> m a
+withTransactionSerializableRunBeamPostgres dbConn = liftIO . withTransactionSerializable dbConn . runBeamPostgres dbConn